+++ src-bdpimport/src/main/java/com/bosch/osmi/sw360/bdp/datasource/BdpApiAccessWrapperNew.java
@@ -13,7 +13,7 @@ import com.bosch.osmi.bdp.access.api.BdpApiAccess;
 import com.bosch.osmi.bdp.access.api.model.ProjectInfo;
 import com.bosch.osmi.bdp.access.impl.BdpApiAccessImpl;
 import com.bosch.osmi.bdp.access.impl.model.ProjectInfoEager;
-import org.eclipse.sw360.datahandler.thrift.bdpimport.RemoteCredentials;
+import org.eclipse.sw360.datahandler.thrift.projectimport.RemoteCredentials;
 
 import java.util.Collection;
